refactor: split Sun defaults setup owners
