+++ service-worker.js
@@ -519,7 +519,10 @@ const APP_SHELL = [
   '/js/sun-context-session-tools.js',
   '/js/sun-correlations.js',
   '/js/sun-defaults.js',
+  '/js/sun-defaults-model.js',
   '/js/sun-defaults-runtime.js',
+  '/js/sun-defaults-setup-renderer.js',
+  '/js/sun-defaults-setup-ui.js',
   '/js/sun-onboarding-ai.js',
   '/js/sun-spectrum-actions.js',
   '/js/sun-spectrum-device.js',
