+++ tests/test-correctness-phase2.js
@@ -191,6 +191,9 @@ const pwaAppShellAssets = [
   '/js/sun-sessions-store.js',
   '/js/sun-session-ui.js',
   '/js/sun-session-actions.js',
+  '/js/sun-defaults-model.js',
+  '/js/sun-defaults-setup-renderer.js',
+  '/js/sun-defaults-setup-ui.js',
   '/js/sun-spectrum-actions.js',
   '/js/sun-spectrum-device.js',
   '/js/sun-spectrum.js',
