+++ tests/test-quality-guardrails.js
@@ -260,7 +260,10 @@ const domainUiCheckJsModules = [
   'js/sun-context-runtime.js',
   'js/sun-context-session-tools.js',
   'js/sun-defaults.js',
+  'js/sun-defaults-model.js',
   'js/sun-defaults-runtime.js',
+  'js/sun-defaults-setup-renderer.js',
+  'js/sun-defaults-setup-ui.js',
   'js/sun-location.js',
   'js/sun-runtime.js',
   'js/sun-spectrum-actions.js',
