+++ tests/verify-modules.js
@@ -112,7 +112,10 @@
     '/js/sun-spectrum-actions.js',
     '/js/sun-spectrum-device.js',
     '/js/sun-runtime.js',
+    '/js/sun-defaults-model.js',
     '/js/sun-defaults-runtime.js',
+    '/js/sun-defaults-setup-renderer.js',
+    '/js/sun-defaults-setup-ui.js',
     '/js/sun-body-silhouette-runtime.js',
     '/js/sun-channel-metrics.js',
     '/js/sun-context-hooks.js',
