+++ tsconfig.json
@@ -298,7 +298,10 @@
     "js/sun-context-session-tools.js",
     "js/sun-correlations.js",
     "js/sun-defaults.js",
+    "js/sun-defaults-model.js",
     "js/sun-defaults-runtime.js",
+    "js/sun-defaults-setup-renderer.js",
+    "js/sun-defaults-setup-ui.js",
     "js/sun-onboarding-ai.js",
     "js/sun-location.js",
     "js/sun-runtime.js",
